BUGFIX: ExtraLazyPersistentCollection uses entity identifiers as default ordering
Cursor based pagination relies on adding every order attribute to the
payload. If custom sorting applies there's a chance of having additional
entities on the following page that have the exact same subset of data
as the last seen item on the current page has. Those following items
will be filtered and thereby not shown unless we add an actual unique
and sorted by attribute. That's by definition the identifier.