BUGFIX: Getter methods changed type detection from other properties
Every getter method annotated with "ExposeProperty" changed the detected
type value of the last actual property of a certain class. This could
have several effects, like switching a certain type from attribute to
relationship or vice versa, or using the wrong class type when searching
for a Sorting or Filter class.