feat: Avoid xand composite clause with only one child expression
The extra lazy collection used to wrap its where clause in XAND
expressions for every subsequent "getCriteria()", which resulted
in a couple of layers of AND calculations, usually only having a
single EQUALS comparison.

This change only wraps multipe WHERE expressions into an XOR
cluase and uses the existing clause "as is" if there's only one
condition.