feat: Rename "request" arguments in Configuration
The "Configuration" object has data like "package key" or "Controller
name", which are meant to point to the set of actions for "list",
"show", "create" and "delete". Those are basically CRUD verbs and rarely
differ between projects. They clearly point to the "Generic" stuff this
package aims to provide.

On the other hand, the "Configuration" object has data like "type name"
and "attributes to be api exposed", which isn't "generic" at all but
depends on the actual domain model.

This change now renames the "generic" arguments and prefixes them with
"request". So "controllerName" becomes "requestControllerName".

When everything not prefixed with "request" being part of the current
domain, it becomes clear what "className" actually means. The class name
of the entity, that is.