fix: ExposableTypeMap::collectKnownTypes() cannot be CompileStatic
The @Flow\CompileStatic annotation calls the annotated method not with
the regular object manager but with the compile time object manager,
which does not make use of proxy classes but public setter methods
such as "inject.*" and "set.*".

Using either @Flow\Inject or constrctor based injection, on the other
hand relies on proxy classes that can just write protected properties
directly, so there is no such setting accessor method.

This means a method with @Flow\CompileStatic cannot be used together
with @Flow\Inject.