fix: Prevent show requests from validating known objects
If a single object is to be displayed which is referenced by its string
primary key, there's no reason for it to pass through the validation
pipeline.

This primarily prevents show requests from loading arbitrary properties
that might not be API-exposed and not even meant to be accessed at all
in a web request, hence improving memory optimization — such as, for
example, large collections.